Be clear more clear about results of enablePersistence()
The code sample here suggests that the developer should use the promise returned from enablePersistence() to get a hold of a new Firestore instance when the promise resolves successfully.  This isn't a requirement, as subsequent queries are supposed to be enqueued to execute after enablePersistence is complete, in either success or failure cases.