Update email.js
The promises for `signInWithEmailAndPassword` and `createUserWithEmailAndPassword` resolve with a `UserCredential` and not a user. Calling the parameter `user` is a left-over from the old (pre-5.0) SDK where it was an actual `FirebaseUser` object.

I renamed the parameter and added an example of how to access the UID, which is the most common action for developers after creating an account/signing in.